feat(oidc-client): enhance service worker flexibility
