+++ packages/oidc-client/src/types.ts
@@ -3,6 +3,7 @@ export type Fetch = typeof window.fetch;
 export type LogoutToken = 'access_token' | 'refresh_token';
 
 export type ServiceWorkerUpdateRequireCallback = (registration:any, stopKeepAlive:Function) => Promise<void>;
+export type ServiceWorkerActivate = () => boolean;
 
 export type OidcConfiguration = {
     client_id: string;
@@ -18,6 +19,8 @@ export type OidcConfiguration = {
     refresh_time_before_tokens_expiration_in_second?: number;
     token_request_timeout?: number;
     service_worker_relative_url?:string;
+    service_worker_keep_alive_path?:string;
+    service_worker_activate?:ServiceWorkerActivate;
     service_worker_only?:boolean;
     service_worker_convert_all_requests_to_cors?:boolean;
     service_worker_update_require_callback?:ServiceWorkerUpdateRequireCallback;
